album gesture refactoring 1차, search error, delete error fix
